Fixed naming schema of unified Near Cache tests to *NearCacheBasicTest
